+++ library/std/src/sys/thread_local/static_local.rs
